Added zephyr -- OpenAI-compatible model (#81)
# Summary

Zephyr is a fine-tune of mistral-7b, optimized for chat. 

# Testing this

Deployed via `truss push`, and did inference with `curl`:


```
$ curl -X POST https://model-rwnpdl13.api.baseten.co/production/predict \
  -H 'Authorization: Api-Key 6EwFTIZh.BOm20EIaOXVob7X7zSXjp5yXXenqWNng' \
  -d '{"messages": [{"role": "user", "content": "What is the mistral wind?"}]}'
"<|user|>\nWhat is the mistral wind?</s> \n<|assistant|>\nThe mistral is a strong and cold northwesterly wind that blows across the Mediterranean Sea and into southern France, Spain, Italy, and Corsica. It typically occurs in the winter months and can reach speeds of up to 100 km/h (60 mph). The mistral is known for its distinctive sound, which resembles a howl, and its powerful and sometimes destructive force. It can be especially dangerous for sailors, as the sudden and violent gusts can cause capsizing and damage to boats.</s>"%
```

---------

Co-authored-by: Bola Malek <bola@baseten.co>
Co-authored-by: Timur Abishev <timur.abishev@baseten.co>
Co-authored-by: joostinyi <63941848+joostinyi@users.noreply.github.com>