Support building trt engines on the fly (#92)
In benchmarking we test a lot of configurations and cold start time
isn't important. To simplify development / improve velocity I've added
trt-llm truss which supports building engines on the fly.

To do so instead of specifying `engine_repository` `engine_build` needs
to be specified. Here's an example of such configuration:
```
engine_build:
  cmd: examples/llama/build.py
  args: --remove_input_padding --use_gpt_attention_plugin float16 --enable_context_fmha --use_gemm_plugin float16 --max_batch_size 64 --use_inflight_batching --max_input_len 2000 --max_output_len 2000 --paged_kv_cache
```
This goes under `model_metadata`, at the same place we put
`engine_repository` currently.
For this configuration it takes about 3 minutes on A100 to load the
model.