Add repo_id to llama repos (#163)
# Summary

In https://github.com/basetenlabs/baseten/pull/7807 & some previous PRs,
we added some validation logic to check on baseten when deploying a
model whether the user has access to the repo in question or not (with
their access token). In order to do this check, we need to have the
repo_id that the model is pointing to in Huggingface in the config (it
has to be in some sort of structured data to read it).

In this PR, we add the correct repo_id to each of the llama repos so we
can validate them.

# Follow-up

A good follow-up to this would be reading the config value in the
model.py files instead of hard-coding them there.