Standardize engine builder configs ahead of model library deployment (#410)
There are a few simultaneous efforts going on here, so I want to be
clear about the scope of this PR.

After Michael's changes to standardize completions and chatcompetions
support in Briton are fully baked in prod, I will re-build the engines
in the model library to be OpenAI compatible and incorporate all of our
latest context builder, tensorrt-llm version, and engine builder
upgrades.

This PR modifies the relevant files to rebuild the engines that are
currently in the library, as well as adds a couple of missing ones.

Separately, there is an effort to re-organize the truss-examples repo.
This PR does not touch that effort, it modifies files in place ahead of
model library deployment.